refactor: extract SetupWizard and CopyButton components, fix setTimeout cleanup
